% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{generate_downloadButton}
\alias{generate_downloadButton}
\title{Download button}
\usage{
generate_downloadButton(id, plot_out, device)
}
\arguments{
\item{id}{Id name of tabset panel.}

\item{plot_out}{Reactive. Plot to save.}

\item{device}{Parameter from \code{\link[ggplot2]{ggsave}}.}
}
\value{
\code{generate_downloadButton} returns output of \code{downloadHandler} for given plot and device.
}
\description{
Generates download button.
}
\details{

}
