% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare HDX data}
\usage{
prepare_data(input, plt = "comparison")
}
\arguments{
\item{input}{List}

\item{plt}{Plot type. Accepts either \code{'comparison'} or \code{'bar'}. Default \code{'comparison'}}
}
\description{
Prepares data.
}
